JetSpectrum: add comparison with ML paper

+++ Jets/JetSpectrum_ResponseMatrixFunctions.C
@@ -422,7 +422,7 @@ void Get_PtResponseMatrix_detectorResponse(TH2D* &H2D_jetPtResponseMatrix_detect
     if (mcIsWeighted) {
       H2D_response->Scale(1./GetNEventsSelected_JetFramework_weighted(file_O2Analysis_MCfileForMatrix, analysisWorkflowMC));
     } else {
-      int Nevents = GetNEventsSelected_JetFramework(file_O2Analysis_MCfileForMatrix, analysisWorkflowMC);
+      double Nevents = GetNEventsSelected_JetFramework(file_O2Analysis_MCfileForMatrix, analysisWorkflowMC);
       for (int iBinX = 0; iBinX < H2D_response->GetNbinsX(); iBinX++) {
         for (int iBinY = 0; iBinY < H2D_response->GetNbinsY(); iBinY++) {
           H2D_response->SetBinContent(iBinX, iBinY, H2D_response->GetBinContent(iBinX, iBinY) * 1./Nevents);
